--- a/src/PermutationGroups.jl
+++ b/src/PermutationGroups.jl
@@ -17,12 +17,14 @@ export firstmoved, fixes, fixedpoints, lastmoved, nfixedpoints
 export base, representative, schreier_sims, sgs, sift
 
 include("types.jl")
-include("utils.jl")
 include("orbit.jl")
 include("words.jl")
-include("schreier.jl")
+include("group_interface.jl")
 include("groups.jl")
+include("schreier.jl")
 include("stabchain.jl")
 include("schreier-sims.jl")
+
+include("utils.jl")
 
 end # module