--- a/src/schreier-sims.jl
+++ b/src/schreier-sims.jl
@@ -49,10 +49,10 @@ function schreier_sims!(sc::StabilizerChain)
 end
 
 @doc doc"""
-    schreier_sims(gens::Vector{perm}, B::Vector{<:Integer}) → StabilizerChain
+    schreier_sims(gens::Vector{perm}, B::AbstractVector{<:Integer}) → StabilizerChain
 Complete `gens` to strong generator set by including `B` as (partial) base.
 """
-function schreier_sims(gens::AbstractVector{Perm{I}}, B::AbstractVector{I}=I[]) where I<:Integer
+function schreier_sims(gens::AbstractVector{<:AbstractPerm}, B::AbstractVector{<:Integer}=Int[]) where I<:Integer
     sc = StabilizerChain(gens, B)
     schreier_sims!(sc)
     return sc.base, sc.sgs, sc.transversals