… .github/workflows/ci.yml
… Project.toml
… src/utils.jl
… src/words.jl
… test/benchmark.jl
… test/groups_interface.jl
… test/orbit.jl
… test/runtests.jl
… test/schreier.jl
… test/stabchain.jl