@kalmarek 2021-04-21
this branch:
```julia
     Testing Running tests...
[ Info: Native iteration over S8 group:
  1.372 ms (80644 allocations: 7.38 MiB)
[ Info: Iteration over K ≅ S8 PermGroup:
  18.098 ms (161282 allocations: 11.69 MiB)
[ Info: Schreier-Sims for Rubik cube-4 group:
  96.380 μs (2099 allocations: 225.72 KiB)
[ Info: Schreier-Sims for Rubik cube-9 group:
  45.131 ms (611136 allocations: 129.55 MiB)
[ Info: Schreier-Sims for SL(4,7):
  220.755 ms (383826 allocations: 444.93 MiB)
[ Info: Schreier-Sims for a direct-product group:
  36.098 ms (53785 allocations: 62.93 MiB)
  2.265 s (769925 allocations: 652.02 MiB)
Test Summary:     | Pass  Total
PermutationGroups |  462    462
     Testing PermutationGroups tests passed 
```

master branch:
```
[ Info: Native iteration over S8 group:
  10.413 ms (201604 allocations: 23.38 MiB)
[ Info: Iteration over K ≅ S8 PermGroup:
  39.925 ms (737152 allocations: 51.92 MiB)
[ Info: Schreier-Sims for Rubik cube-4 group:
  90.584 μs (2065 allocations: 224.62 KiB)
[ Info: Schreier-Sims for Rubik cube-9 group:
  43.227 ms (611100 allocations: 129.55 MiB)
[ Info: Schreier-Sims for SL(4,7):
  211.704 ms (383804 allocations: 444.93 MiB)
[ Info: Schreier-Sims for a direct-product group:
  33.654 ms (53759 allocations: 62.93 MiB)
  2.752 s (3543438 allocations: 1.43 GiB)
```
* naive iteration is due to new AA (and me optimizing it for 2 allocs/perm)
* I made 4 allocs/perm in iteration over `PermGroup` (2 for perm, 2 for Permutation). much better product iterator could be devised (the current one takes product of all elements from transversal each time), making it 0-allocation O(n·k²), but O(n·k) should be possible (n=maximal orbit length, k=number of orbits). I expect the first iteration could be of ~3-4 times faster.
* a consistent (small) slowdown in Schreier-Sims is probably due to `Permutation` layer on top. I'd need to investigate.